backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
supabase>=2.28.2
ultralytics==8.3.0
python-dotenv==1.0.1
pydantic==2.8.2
haversine==2.8.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
